% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{get_weather_forecast}
\alias{get_weather_forecast}
\title{Weather forecast API}
\usage{
get_weather_forecast(city, date)
}
\arguments{
\item{city}{A string of a city with an uppercase for the first letter}

\item{date}{A string of a date in format "YYYY-MM-DD"}
}
\value{
A \code{data.frame} with time, temperature, humidity, precipitation, cloudcover, wind speed, wind direction and city for every hour.
}
\description{
Weather forecast data for five large cities in Sweden. Could be used with a shiny application.
}
\examples{
# Weather forecast for Stockholm today.
get_weather_forecast(city="Stockholm", date=Sys.Date()) 

# Weather forecast for Göteborg tomorrow.
get_weather_forecast(city="Göteborg", date=Sys.Date()+1)

}
\references{
Open-Meteo. Weather Forecast API. \url{https://open-meteo.com/en/docs}
}
