% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pack_disc.R
\docType{package}
\name{Weather4cats}
\alias{Weather4cats}
\title{Weather forecast API}
\description{
Weather forecast data for five large cities in Sweden. Could be used with a shiny application.
}
\author{
Leonard Persson Norblad \email{leope744@student.liu.se} Jacob Sibanda \email{jacsi259@student.liu.se}
}
